Move file with similar name L1TMuonEndcapParamsRcd.*
In the directory is a file with the name L1TMuonEndCapParamsRcd.*
where that file is actually used. We can't get rid of the old
file as the DB needs that class to exist.
