+++ CondFormats/DataRecord/src/L1TMuonEndcapParamsRcd_deprecated.cc
@@ -6,7 +6,7 @@
 
 // automatically generated by make_records.pl
 
-#include "CondFormats/DataRecord/interface/L1TMuonEndcapParamsRcd.h"
+#include "CondFormats/DataRecord/interface/L1TMuonEndcapParamsRcd_deprecated.h"
 #include "FWCore/Framework/interface/eventsetuprecord_registration_macro.h"
 
 EVENTSETUP_RECORD_REG(L1TMuonEndcapParamsRcd);
